Abraham a develop - 01/09/2021

+++ database/migrations/2021_06_10_224007_create_inversions_table.php
@@ -20,9 +20,9 @@ class CreateInversionsTable extends Migration
             $table->bigInteger('orden_id')->unsigned();
             $table->foreign('orden_id')->references('id')->on('orden_purchases')->onUpdate('cascade')->onDelete('cascade');
             $table->double('invertido');
-            $table->double('ganacia');
+            $table->integer('ganacia');
             $table->tinyInteger('status')->default(1)->comment('1 - activo , 2 - culminada');
-            $table->double('max_ganancia')->nullable();
+            $table->integer('max_ganancia')->nullable();
             $table->timestamps();
         });
     }
